add tag link and contact part

+++ src/main/java/com/HYX/webDev/dao/LinkMapper.java
@@ -25,6 +25,6 @@ public interface LinkMapper {
 
     int getTotalLink(Map param);
 
-    int deleteBatch(Long[] ids);
+    int deleteBatch(Integer[] ids);
 
 }
